階段性任務(譜曲功能)39子項未完成

+++ free_marker/project-yueh-pu/web/src/component/personalRhythm/index.js
@@ -1,3 +1,5 @@
+const edit = true;
+
 import {inject} from "mobx-react";
 import BasePersonalRhythmComponent from "./BasePersonalRhythmComponent";
 import {
@@ -43,9 +45,12 @@ class PersonalRhythmComponent extends BasePersonalRhythmComponent {
         const self = this;
         return async () => {
             const favoritePu = param.object;
-            favoritePu.deleteFavoritePuItem(self).then((result) => {
-                this.showInfoSnackMessage(`已成功刪除「${favoritePu.name}」`)
-            })
+            if (self.getStore().hasOnClickOfDeleteOfMenuItem())
+                await self.getStore().getOnClickOfDeleteOfMenuItem(favoritePu);
+            else
+                favoritePu.deleteFavoritePuItem(self).then((result) => {
+                    this.showInfoSnackMessage(`已成功刪除「${favoritePu.name}」`)
+                })
         }
     }
